+++ pyprom/dataload.py
@@ -87,12 +87,11 @@ class GDALLoader(Loader):
             target.ImportFromEPSG(epsg_code)
             transform = osr.CoordinateTransformation(spatialRef, target)
             # create a reverse transform for translating back
-            #  into Native GDAL coordiantes
+            #  into Native GDAL coordinates
             reverse_transform = osr.CoordinateTransformation(target,
                                                              spatialRef)
             self.linear_unit = spatialRef.GetLinearUnits()
             self.linear_unit_name = spatialRef.GetLinearUnitsName()
-
 
             trimmed_file_name = filename.split("/")[-1]
             # Create out DataMap Object.